refactor: デプロイ用にトップページを元のままに維持する。
