+++ youkai-front/nuxt.config.js
@@ -59,7 +59,8 @@ export default {
   axios: {
     baseURL: process.env.NODE_ENV === "production" ? "https://youkaiosugi.com/" : "http://localhost:8080/",
     // prefix: '/api',  // FIXME なぜか「効かない」ので使わない
-    debug: process.env.NODE_ENV !== "production",
+    // debug: process.env.NODE_ENV !== "production",
+    debug: true,  // FIXME 本番での挙動を確認するため。
   },
 
   // PWA module configuration: https://go.nuxtjs.dev/pwa
