+++ youkai-front/src/layouts/default.vue
@@ -85,15 +85,20 @@ export default class extends Vue {
   selfVersion = "";
 
   private readonly items = [
-    {
-      icon: 'mdi-apps',
-      title: '発見済パスワード検索',
-      to: '/',
-    },
+    // {
+    //   icon: 'mdi-apps',
+    //   title: '発見済パスワード検索',
+    //   to: '/',
+    // },
+    // {
+    //   icon: 'mdi-apps',
+    //   title: '総当りチャレンジ',
+    //   to: '/challenge',
+    // },
     {
       icon: 'mdi-apps',
       title: '総当りチャレンジ',
-      to: '/challenge',
+      to: '/',
     },
     {
       icon: 'mdi-apps',
